requirements.txt
xarray
netCDF4
hvplot
panel
pandas
ncplot
dill
python-dateutil
matplotlib
jupytext
nctoolkit>=1.1.1
jellyfish
geopandas
holoviews
plotnine
jupyter-book
seawater
cmocean
numpydoc
sphinx_rtd_theme
nbsphinx
molmass


